actor_effects

Overview of statistics in the actor-oriented model, see Details.

Usage

actor_effects(step = NULL)

Arguments

step

outputs all statistics in the sender activity step (if `step = sender`) or receiver choice step (if `step = receiver`).

Value

Returns a list of available effects and their corresponding statistics based on the specified `step` (sender or receiver).

Details

Overview of statistics in the actor-oriented model.

A list of available effects and their corresponding statistics for the sender activity rate step:

baseline()

send()

indegreeSender()

outdegreeSender()

totaldegreeSender()

recencySendSender()

recencyReceiveSender()

psABA()

psABB()

psABX()

A list of available effects and their corresponding statistics for the receiver choice step:

receive()

tie()

same()

difference()

average()

indegreeReceiver()

outdegreeReceiver()

totaldegreeReceiver()

inertia()

reciprocity()

otp()

itp()

osp()

isp()

rrankSend()

rrankReceive()

recencySendReceiver()

recencyReceiveReceiver()

recencyContinue()

psABAB()

psABBA()

psABXA()

psABXB()

psABAY()

psABBY()

psABXY()

Examples

# List of available effects for both the sender and receiver step
actor_effects()
#> $sender
#>  [1] "baseline"             "send"                 "indegreeSender"      
#>  [4] "outdegreeSender"      "totaldegreeSender"    "recencySendSender"   
#>  [7] "recencyReceiveSender" "psABA"                "psABB"               
#> [10] "psABX"                "userStat"            
#> 
#> $receiver
#>  [1] "receive"                "same"                   "difference"            
#>  [4] "average"                "tie"                    "inertia"               
#>  [7] "reciprocity"            "indegreeReceiver"       "outdegreeReceiver"     
#> [10] "totaldegreeReceiver"    "otp"                    "itp"                   
#> [13] "osp"                    "isp"                    "rrankSend"             
#> [16] "rrankReceive"           "recencySendReceiver"    "recencyReceiveReceiver"
#> [19] "recencyContinue"        "psABBA"                 "psABAB"                
#> [22] "psABXB"                 "psABXA"                 "psABAY"                
#> [25] "psABBY"                 "psABXY"                 "userStat"              
#> 

# List of available effects for the sender step
actor_effects(step = "sender")
#>  [1] "baseline"             "send"                 "indegreeSender"      
#>  [4] "outdegreeSender"      "totaldegreeSender"    "recencySendSender"   
#>  [7] "recencyReceiveSender" "psABA"                "psABB"               
#> [10] "psABX"                "userStat"            

# List of available effects for the receiver step
actor_effects(step = "receiver")
#>  [1] "receive"                "same"                   "difference"            
#>  [4] "average"                "tie"                    "inertia"               
#>  [7] "reciprocity"            "indegreeReceiver"       "outdegreeReceiver"     
#> [10] "totaldegreeReceiver"    "otp"                    "itp"                   
#> [13] "osp"                    "isp"                    "rrankSend"             
#> [16] "rrankReceive"           "recencySendReceiver"    "recencyReceiveReceiver"
#> [19] "recencyContinue"        "psABBA"                 "psABAB"                
#> [22] "psABXB"                 "psABXA"                 "psABAY"                
#> [25] "psABBY"                 "psABXY"                 "userStat"