tie_effects

Overview of statistics in the tie-oriented model, see Details.

Usage

tie_effects(directed = NULL, endogenous = NULL)

Arguments

directed

logical value. The function outputs all statistics in the tie-oriented model for directed events if true, or all statistics in the tie-oriented model for undirected events if false.

endogenous

logical value. The function outputs all endogenous statistics in the tie-oriented model if true, or all exogenous statistics if false

Value

Returns a list of available effects and their corresponding statistics.

Details

Overview of statistics in the tie-oriented model.

Baseline:

baseline

Exogenous statistics:

send()

receive()

tie()

same()

difference()

average()

minimum()

maximum()

event()

userStat()

Endogenous statistics:

indegreeSender()

indegreeReceiver()

outdegreeSender()

outdegreeReceiver()

totaldegreeSender()

totaldegreeReceiver()

totaldegreeDyad()

degreeMin()

degreeMax()

degreeDiff()

inertia()

reciprocity()

otp()

itp()

osp()

isp()

sp()

psABBA()

psABBY()

psABXA()

psABXB()

psABXY()

psABAY()

psABAB()

rrankSend()

rrankReceive()

recencySendSender()

recencySendReceiver()

recencyReceiveSender()

recencyReceiveReceiver()

recencyContinue()

FEtype()

Examples

# List of available effects 
tie_effects()
#>  [1] "baseline"               "FEtype"                 "send"                  
#>  [4] "receive"                "tie"                    "same"                  
#>  [7] "difference"             "average"                "minimum"               
#> [10] "maximum"                "event"                  "inertia"               
#> [13] "reciprocity"            "indegreeSender"         "indegreeReceiver"      
#> [16] "outdegreeSender"        "outdegreeReceiver"      "totaldegreeSender"     
#> [19] "totaldegreeReceiver"    "totaldegreeDyad"        "degreeMin"             
#> [22] "degreeMax"              "degreeDiff"             "otp"                   
#> [25] "itp"                    "osp"                    "isp"                   
#> [28] "sp"                     "psABBA"                 "psABBY"                
#> [31] "psABXA"                 "psABXB"                 "psABXY"                
#> [34] "psABAY"                 "psABAB"                 "rrankSend"             
#> [37] "rrankReceive"           "recencyContinue"        "recencySendSender"     
#> [40] "recencySendReceiver"    "recencyReceiveSender"   "recencyReceiveReceiver"
#> [43] "userStat"              

# List of available effects for undirected networks
tie_effects(directed = FALSE)
#>  [1] "baseline"        "FEtype"          "tie"             "same"           
#>  [5] "difference"      "average"         "minimum"         "maximum"        
#>  [9] "event"           "inertia"         "totaldegreeDyad" "degreeMin"      
#> [13] "degreeMax"       "degreeDiff"      "sp"              "psABAY"         
#> [17] "psABAB"          "recencyContinue" "userStat"       

# List of available endogenous effects for undirected networks
tie_effects(directed = FALSE, endogenous = TRUE)
#>  [1] "baseline"        "FEtype"          "inertia"         "totaldegreeDyad"
#>  [5] "degreeMin"       "degreeMax"       "degreeDiff"      "sp"             
#>  [9] "psABAY"          "psABAB"          "recencyContinue"